Added the beginning of a Tree data structure.
Using a fluent type builder strategy child can be added consecutively in a chain, siblings can also be added.

Addition of an ancestor sibling is being worked on to enable backtracking up the tree to a certain node before adding another sibling and subsequent children to create a new branch of the tree.