Unit tests added for Tree datastructure.
Equals method and tests added.

Addition of an ancestor sibling method to enable backtracking up the tree to a certain node before adding another sibling and subsequent children to create a new branch of the tree.

Default call without depth argument on ancestor sibling will only go up to the grandparent node to add a child, in other words it will add a sibling to the parent of the current child node.